# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "User API",
    "version": "1.0.0",
    "description": "API ini menyediakan serangkaian endpoint untuk mengelola users, accounts dan transactions dalam sistem. Fungsionalitas yang ditawarkan meliputi post,get,put,dan delete. API ini dirancang untuk mendukung aplikasi berbasis web dan seluler, memastikan pengalaman pengguna yang aman dan efisien.  \"berjalan di localhost:3000\""
  },
  "servers": [
    {
      "url": "http://localhost:3000/",
      "description": "Server Localhost"
    }
  ],
...
  "paths": {
...
    "/api/v1/accounts": {
      "post": {
        "tags": [
          "Accounts"
        ],
        "summary": "Tambah Account",
        "operationId": "addAccount",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewAccount"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Akun berhasil ditambahkan",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddAccountResponse"
                }
              }
            }
          },
          "400": {
            "description": "Nomor bank ini sudah dipakai untuk bank yang sama",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorAccount400"
                }
              }
            }
          },
          "404": {
            "description": "User ID tidak ditemukan",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorUser404"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Accounts"
        ],
        "summary": "Dapatkan Daftar Account",
        "operationId": "getAccounts",
        "responses": {
          "200": {
            "description": "Berhasil menampilkan daftar akun",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetAccountsResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/accounts/{accountId}": {
      "parameters": [
        {
          "name": "accountId",
          "in": "path",
          "required": true,
          "description": "ID",
          "schema": {
            "type": "integer"
          }
        }
      ],
      "get": {
        "tags": [
          "Accounts"
        ],
        "summary": "Dapatkan Detail Account",
        "operationId": "getAccountById",
        "responses": {
          "200": {
            "description": "Berhasil menampilkan detail akun",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetAccountResponse"
                }
              }
            }
          },
          "404": {
            "description": "Akun tidak ditemukan",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorAccount404"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Accounts"
        ],
        "summary": "Update Account",
        "operationId": "updateAccount",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewAccountPut"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Akun berhasil diupdate",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateAccountResponse"
                }
              }
            }
          },
          "404": {
            "description": "Akun tidak ditemukan",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorAccount404"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Accounts"
        ],
        "summary": "Hapus Account",
        "operationId": "deleteAccount",
        "responses": {
          "200": {
            "description": "Akun berhasil dihapus",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteAccountResponse"
                }
              }
            }
          },
          "404": {
            "description": "Akun tidak ditemukan",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorAccount404"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/transactions/{transactionId}": {
      "parameters": [
        {
          "name": "transactionId",
          "in": "path",
          "required": true,
          "description": "ID",
          "schema": {
            "type": "integer"
          }
        }
      ],
      "get": {
        "tags": [
          "Transactions"
        ],
        "summary": "Dapatkan Detail Transaksi",
        "operationId": "getTransactionById",
        "responses": {
          "200": {
            "description": "Berhasil menampilkan detail transaksi",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetTransactionIdResponse"
                }
              }
            }
          },
          "404": {
            "description": "Transaksi tidak ditemukan",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorTransaction404"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "AuthHeader": {
        "type": "http",
        "scheme": "bearer",
        "description": "Gunakan header `Authorization` dengan format `Bearer <token>` untuk autentikasi anda bisa memasukkan token di sini yang sudah di buat pada melalui login.\n"
      }
    },
    "schemas": {
...
      "ErrorUser404": {
        "type": "object",
        "properties": {
          "status": {
            "type": "integer",
            "example": 404
          },
          "message": {
            "type": "string",
            "example": "tidak ada user"
          }
        }
      },
      "NewAccount": {
        "type": "object",
        "properties": {
          "user_id": {
            "type": "integer",
            "example": 1
          },
          "bank_name": {
            "type": "string",
            "example": "Bank A"
          },
          "bank_account_number": {
            "type": "string",
            "example": "1234567890"
          },
          "balance": {
            "type": "number",
            "example": 1000000
          }
        },
        "required": [
          "id",
          "user_id",
          "bank_name",
          "bank_account_number",
          "balance"
        ]
      },
      "NewAccountPut": {
        "type": "object",
        "properties": {
          "user_id": {
            "type": "integer",
            "example": 1
          },
          "bank_name": {
            "type": "string",
            "example": "Bank B"
          },
          "bank_account_number": {
            "type": "string",
            "example": "1234567890"
          },
          "balance": {
            "type": "number",
            "example": 1000000
          }
        }
      },
      "AddAccountResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "integer",
            "example": 201
          },
          "message": {
            "type": "string",
            "example": "akun berhasil ditambahkan"
          },
          "data": {
            "properties": {
              "id": {
                "type": "integer",
                "example": 1
              },
              "user_id": {
                "type": "integer",
                "example": 1
              },
              "bank_name": {
                "type": "string",
                "example": "Bank A"
              },
              "bank_account_number": {
                "type": "string",
                "example": "1234567890"
              },
              "balance": {
                "type": "number",
                "example": 1000000
              }
            }
          }
        }
      },
      "GetAccountsResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "integer",
            "example": 200
          },
          "message": {
            "type": "string",
            "example": "berhasil menampilkan daftar akun"
          },
          "data": {
            "properties": {
              "id": {
                "type": "integer",
                "example": 1
              },
              "user_id": {
                "type": "integer",
                "example": 1
              },
              "bank_name": {
                "type": "string",
                "example": "Bank A"
              },
              "bank_account_number": {
                "type": "string",
                "example": "1234567890"
              },
              "balance": {
                "type": "number",
                "example": 1000000
              }
            }
          }
        }
      },
      "GetAccountResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "integer",
            "example": 200
          },
          "message": {
            "type": "string",
            "example": "berhasil menampilkan detail akun"
          },
          "data": {
            "properties": {
              "id": {
                "type": "integer",
                "example": 1
              },
              "user_id": {
                "type": "integer",
                "example": 1
              },
              "bank_name": {
                "type": "string",
                "example": "Bank A"
              },
              "bank_account_number": {
                "type": "string",
                "example": "1234567890"
              },
              "balance": {
                "type": "number",
                "example": 1000000
              }
            }
          }
        }
      },
      "UpdateAccountResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "integer",
            "example": 200
          },
          "message": {
            "type": "string",
            "example": "akun berhasil diupdate"
          },
          "data": {
            "properties": {
              "id": {
                "type": "integer",
                "example": 1
              },
              "user_id": {
                "type": "integer",
                "example": 1
              },
              "bank_name": {
                "type": "string",
                "example": "Bank B"
              },
              "bank_account_number": {
                "type": "string",
                "example": "1234567890"
              },
              "balance": {
                "type": "number",
                "example": 1000000
              }
            }
          }
        }
      },
      "DeleteAccountResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "integer",
            "example": 200
          },
          "message": {
            "type": "string",
            "example": "akun berhasil dihapus"
          }
        }
      },
      "ErrorAccount400": {
        "type": "object",
        "properties": {
          "status": {
            "type": "integer",
            "example": 400
          },
          "message": {
            "type": "string",
            "example": "nomor bank ini sudah dipakai untuk bank yang sama"
          }
        }
      },
      "ErrorAccount404": {
        "type": "object",
        "properties": {
          "status": {
            "type": "integer",
            "example": 404
          },
          "message": {
            "type": "string",
            "example": "tidak ada akun"
          }
        }
      },
...
      "GetTransactionIdResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "integer",
            "example": 200
          },
          "message": {
            "type": "string",
            "example": "berhasil menampilkan detail transaksi"
          },
          "data": {
            "type": "object",
            "properties": {
              "id": {
                "type": "integer",
                "example": 1
              },
              "amount": {
                "type": "integer",
                "example": 200000
              },
              "source_account": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "integer",
                    "example": 1
                  },
                  "bank_name": {
                    "type": "string",
                    "example": "Bank A"
                  },
                  "bank_account_number": {
                    "type": "string",
                    "example": "1222222"
                  },
                  "user": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "integer",
                        "example": 1
                      },
                      "name": {
                        "type": "string",
                        "example": "Arnold"
                      },
                      "email": {
                        "type": "string",
                        "example": "y@gmail.com"
                      }
                    }
                  }
                }
              },
              "destination_account": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "integer",
                    "example": 2
                  },
                  "bank_name": {
                    "type": "string",
                    "example": "Bank D"
                  },
                  "bank_account_number": {
                    "type": "string",
                    "example": "15555555"
                  },
                  "user": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "integer",
                        "example": 2
                      },
                      "name": {
                        "type": "string",
                        "example": "Yuki Eslintz"
                      },
                      "email": {
                        "type": "string",
                        "example": "yukicek@gmail.com"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
...
      "ErrorTransaction404": {
        "type": "object",
        "properties": {
          "status": {
            "type": "integer",
            "example": 404
          },
          "message": {
            "type": "string",
            "example": "transaksi tidak ada"
          }
        }
      },
...
    }
  }
}